--- src/main/java/calculator/Operation.java
@@ -1,4 +1,0 @@
-package calculator;
-
-public class Operation {
-}
