+++ src/main/java/calculator/view/OutputView.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.view;
 
 public class OutputView {
 }
